+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+- Disable adjust for brightness when scheduled dark mode on
+
 ## [1.1.5] - 2019-05-02
 ### Added
 - Partial translations to Japanese and Korean